refactor: 디렉토리 이동 및 로또 번호에 보너스 볼이 포함된 경우에 대한 테스트 코드 추가
